requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.29.0
langchain-core==0.3.72
langchain-ollama==0.3.6
pandas==2.3.1
requests==2.32.4
python-dotenv==1.0.1
aiohttp==3.12.15
python-multipart==0.0.20
Pillow==11.3.0
